Hessian conversion/diag_every_n fix
Conversion between Cartesian and internal Hessians were being performed
incorrectly; this has been fixed. I have verified that setting the null
space of the Hessian to be the geometric mean of the absolute
eigenvalues of the non-redundant space improves performance when
diag_every_n > 0, as expected.

Additionally, reset the steps since diagonalization when initializing
Hessian matrix.